+++ app/src/main/java/de/der_nik/wakeup/ui/alarmclock/AlarmClockActivity.kt
@@ -9,6 +9,8 @@ class AlarmClockActivity : AppCompatActivity() {
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
         setContentView(R.layout.alarm_clock_activity)
+        setSupportActionBar(findViewById(R.id.toolbar))
+
         if (savedInstanceState == null) {
             supportFragmentManager.beginTransaction()
                 .replace(R.id.container, AlarmClockFragment.newInstance())
