+++ app/src/main/java/de/der_nik/wakeup/ui/main/alarmlist/AlarmListViewModel.kt
@@ -30,39 +30,39 @@ class AlarmListViewModel(application: Application) : AndroidViewModel(applicatio
     private val scope = CoroutineScope(coroutineContext)
 
     init {
-        val wakeUpDao =WakeUpRoomDatabase.getDatabase(application).wakeUpDao()
-        repository = WakeUpRepository(wakeUpDao)
-        allAlarms = repository.getAlarms()
+            val wakeUpDao =WakeUpRoomDatabase.getDatabase(application).wakeUpDao()
+            repository = WakeUpRepository(wakeUpDao)
+            allAlarms = repository.getAlarms()
     }
 
     fun saveSingle(alarm: LiveData<AlarmTime>) = scope.launch(Dispatchers.IO) {
         repository.insert(alarm.value!!)
     }
 
-    inner class AlarmListItemViewModel: ViewModel(){
-
-        val alarmInfo = MutableLiveData<String>()
-        val alarmDate = MutableLiveData<String>()
-        val alarmActive = MutableLiveData<Boolean>()
-
-        fun onItemClick(alarm: AlarmTime){
-            Log.d("AlarmItem", "Button Clicked: " + alarm.name)
-            val context = getApplication<Application>()
-            Toast.makeText(getApplication(), "Clicked: ${alarm.name}", Toast.LENGTH_LONG).show()
-            val intent = Intent( context, EditAddAlarmActivity::class.java)
-            intent.putExtra("id", alarm.id)
-            context.startActivity(intent)
-        }
-
-        fun ontoggleActiveSwitch(alarm: AlarmTime){
-            val alarmLD = MutableLiveData<AlarmTime>()
-            alarmLD.value = alarm
-            if(!alarmLD.value!!.active){
-                AlarmClockManager.getInstance().activateAlarm(getApplication(), alarmLD.value!!)
-            } else {
-                AlarmClockManager.getInstance().deactivateAlarm(getApplication(), alarmLD.value!!)
-            }
-            saveSingle(alarmLD)
-        }
-    }
+//    inner class AlarmListItemViewModel: ViewModel(){
+//
+//        val alarmInfo = MutableLiveData<String>()
+//        val alarmDate = MutableLiveData<String>()
+//        val alarmActive = MutableLiveData<Boolean>()
+//
+//        fun onItemClick(alarm: AlarmTime){
+//            Log.d("AlarmItem", "Button Clicked: " + alarm.name)
+//            val context = getApplication<Application>()
+//            Toast.makeText(getApplication(), "Clicked: ${alarm.name}", Toast.LENGTH_LONG).show()
+//            val intent = Intent( context, EditAddAlarmActivity::class.java)
+//            intent.putExtra("id", alarm.id)
+//            context.startActivity(intent)
+//        }
+//
+//        fun ontoggleActiveSwitch(alarm: AlarmTime){
+//            val alarmLD = MutableLiveData<AlarmTime>()
+//            alarmLD.value = alarm
+//            if(!alarmLD.value!!.active){
+//                AlarmClockManager.getInstance().activateAlarm(getApplication(), alarmLD.value!!)
+//            } else {
+//                AlarmClockManager.getInstance().deactivateAlarm(getApplication(), alarmLD.value!!)
+//            }
+//            saveSingle(alarmLD)
+//        }
+//    }
 }
