+++ app/src/main/res/layout/alarm_item.xml
@@ -3,7 +3,7 @@
 
     <data>
         <variable name="viewModel"
-                  type="de.der_nik.wakeup.ui.main.alarmlist.AlarmListViewModel.AlarmListItemViewModel"/>
+                  type="de.der_nik.wakeup.ui.main.alarmlist.AlarmListItemViewModel"/>
     </data>
 
     <android.support.v7.widget.CardView
