+++ app/src/main/res/layout/edit_add_alarm_activity.xml
@@ -1,9 +1,19 @@
 <?xml version="1.0" encoding="utf-8"?>
 <layout xmlns:android="http://schemas.android.com/apk/res/android"
         xmlns:tools="http://schemas.android.com/tools"
+        xmlns:app="http://schemas.android.com/apk/res-auto"
         tools:context=".ui.editaddalarm.EditAddAlarmActivity">
-    <android.support.design.widget.CoordinatorLayout android:layout_width="match_parent"
+    <LinearLayout android:orientation="vertical" android:layout_width="match_parent"
                                                      android:layout_height="match_parent">
+
+        <android.support.v7.widget.Toolbar
+                android:id="@+id/toolbar"
+                android:layout_width="match_parent"
+                android:layout_height="?attr/actionBarSize"
+                android:background="?attr/colorPrimary"
+                android:elevation="4dp"
+                android:theme="@style/ThemeOverlay.AppCompat.ActionBar"
+                app:popupTheme="@style/ThemeOverlay.AppCompat.Light"/>
 
 
         <LinearLayout
@@ -14,6 +24,6 @@
 
 
 
-    </android.support.design.widget.CoordinatorLayout>
+    </LinearLayout>
 </layout>
